blueteam/pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "eqmon-blue-team"
version = "0.1.0"
description = "EQMON Blue Team - Defensive Security Monitoring & CMMC Compliance"
requires-python = ">=3.11"
dependencies = [
    "psycopg2-binary>=2.9",
    "click>=8.1",
    "rich>=13.0",
    "pyyaml>=6.0",
    "jinja2>=3.1",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23"]

[project.scripts]
blueteam = "blueteam.cli:main"

[tool.setuptools.packages.find]
include = ["blueteam*"]
